+++ packages/components/src/components.d.ts
@@ -1220,7 +1220,8 @@ export namespace Components {
          */
         "itemsPerPage": number;
         /**
-          * The maximum number of page links rendered.
+          * Has no effect anymore
+          * @deprecated since v3.10.0, will be removed with next major release
          */
         "maxNumberOfPageLinks"?: BreakpointCustomizable<PaginationMaxNumberOfPageLinks>;
         /**
@@ -3778,7 +3779,8 @@ declare namespace LocalJSX {
          */
         "itemsPerPage"?: number;
         /**
-          * The maximum number of page links rendered.
+          * Has no effect anymore
+          * @deprecated since v3.10.0, will be removed with next major release
          */
         "maxNumberOfPageLinks"?: BreakpointCustomizable<PaginationMaxNumberOfPageLinks>;
         /**