+++ packages/components-js/projects/partials/tests/unit/specs/getInitialStyles.spec.ts
@@ -52,16 +52,16 @@ describe('format: jsx', () => {
 describe('format: sha256', () => {
   it('should return hash for core styles', async () => {
     const result = getInitialStyles({ format: 'sha256' });
-    expect(result).toMatchInlineSnapshot(`"'sha256-7W+9h38YfcXwncu6iNGYRieAYr3dSfjPvM3HtFZ9/vs='"`);
+    expect(result).toMatchInlineSnapshot(`"'sha256-ZfnOhvEX2eOryDFV5e2OfWReSZ3vdOp/+4BVMriCr8s='"`);
   });
 
   it('should return hash for core styles with custom prefix', async () => {
     const result = getInitialStyles({ format: 'sha256', prefix: 'custom-prefix' });
-    expect(result).toMatchInlineSnapshot(`"'sha256-UL/ni9ffAJSXII7qvzkY2t/0MIK3ZIfhX13DYWgeQt8='"`);
+    expect(result).toMatchInlineSnapshot(`"'sha256-RXIuVwqtHEAGJHEmCemmf8YFKX3cwaHXViAbeKr0ZW4='"`);
   });
 
   it('should return hash for core styles with multiple custom prefixes', async () => {
     const result = getInitialStyles({ format: 'sha256', prefix: ['', 'some-prefix', 'another-prefix'] });
-    expect(result).toMatchInlineSnapshot(`"'sha256-F8RT4slD88Vv7DYmdaCrSknbyozWivtTUFJL5z0d4t4='"`);
+    expect(result).toMatchInlineSnapshot(`"'sha256-JuCtI7XAfve+WrlX2UfjLoiaB0GMYPNzoSjQu22owp8='"`);
   });
 });