+++ packages/components/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
+- `Flyout Navigation` ([#2906](https://github.com/porsche-design-system/porsche-design-system/pull/2906))
 - Prop `submit-button` to show/hide a submit button for `Text Field Wrapper` `type="search"` if wrapped inside a form
   ([#2908](https://github.com/porsche-design-system/porsche-design-system/pull/2908))
 
@@ -23,7 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Hide `border-bottom` of `Accordion` if used as single item
   ([#2911](https://github.com/porsche-design-system/porsche-design-system/pull/2911))
-- `display` css property can be overridden on all components, e.g. within media queries  
+- `display` css property can be overridden on all components, e.g. to make use of `display: none;` within media
+  queries  
   ([#2913](https://github.com/porsche-design-system/porsche-design-system/pull/2913))
 
 ### [3.9.0] - 2023-11-24