+++ packages/components/CHANGELOG.md
@@ -26,6 +26,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `display` css property can be overridden on all components, e.g. within media queries  
   ([#2913](https://github.com/porsche-design-system/porsche-design-system/pull/2913))
 
+#### Fixed
+
+- `Accordion`: Alignment of slotted heading with custom padding ([2920](https://github.com/porsche-design-system/porsche-design-system/pull/2920))
+
 ### [3.9.0] - 2023-11-24
 
 ### [3.9.0-rc.0] - 2023-11-23