+++ packages/components/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
+- `Link Tile Product` ([#2909](https://github.com/porsche-design-system/porsche-design-system/pull/2909))
 - Prop `submit-button` to show/hide a submit button for `Text Field Wrapper` `type="search"` if wrapped inside a form
   ([#2908](https://github.com/porsche-design-system/porsche-design-system/pull/2908))
 
@@ -28,7 +29,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Pagination`: Prop `maxNumberOfPageLinks` is deprecated and has no effect anymore, instead there is responsive
   behavior out of the box with full SSR support
   ([#2898](https://github.com/porsche-design-system/porsche-design-system/pull/2898))
-
 
 ### [3.9.0] - 2023-11-24