+++ packages/components/CHANGELOG.md
@@ -16,7 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
-- `Link Tile Product` ([#2909](https://github.com/porsche-design-system/porsche-design-system/pull/2909))
+- **[EXPERIMENTAL]** `Link Tile Product`
+  ([#2909](https://github.com/porsche-design-system/porsche-design-system/pull/2909))
 - Prop `submit-button` to show/hide a submit button for `Text Field Wrapper` `type="search"` if wrapped inside a form
   ([#2908](https://github.com/porsche-design-system/porsche-design-system/pull/2908))