+++ packages/components/CHANGELOG.md
@@ -14,8 +14,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+### [3.10.0-rc.0] - 2023-12-07
+
 #### Added
 
+- **[EXPERIMENTAL]** `Flyout Navigation`
+  ([#2906](https://github.com/porsche-design-system/porsche-design-system/pull/2906))
 - Prop `submit-button` to show/hide a submit button for `Text Field Wrapper` `type="search"` if wrapped inside a form
   ([#2908](https://github.com/porsche-design-system/porsche-design-system/pull/2908))
 
@@ -23,7 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - `Accordion`: removed `border-bottom` if used standalone
   ([#2911](https://github.com/porsche-design-system/porsche-design-system/pull/2911))
-- `display` css property can be overridden on all components, e.g. within media queries  
+- `display` css property can be overridden on all components, e.g. to make use of `display: none;` within media
+  queries  
   ([#2913](https://github.com/porsche-design-system/porsche-design-system/pull/2913))
 - `Pagination`: Prop `maxNumberOfPageLinks` is deprecated and has no effect anymore, instead there is responsive
   behavior out of the box with full SSR support